requirements.txt
allure-pytest==2.8.24
allure-python-commons==2.8.24
Appium-Python-Client==1.0.2
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
importlib-metadata==3.1.1
iniconfig==1.1.1
numpy==1.19.4
packaging==20.7
Pillow==8.0.1
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
pytest-repeat==0.9.1
pytest-rerunfailures==9.1.1
PyYAML==5.3.1
selenium==3.141.0
six==1.15.0
toml==0.10.2
urllib3==1.26.2
zipp==3.4.0
tidevice==0.4.3
imageio==2.9.0
facebook-wda==1.4.0
adbutils~=0.11.0
sounddevice==0.4.1
playsound==1.2.2
scipy==1.6.3
openpyxl==3.0.7